refactor: organize SDK conversation hosts
