+++ src/sdk/conversation/index.ts
@@ -1,0 +1,29 @@
+export { ConversationAgentService } from './headless/index.js';
+export type {
+  ConversationAgentCredentialContext,
+  ConversationAgentCredentialPreflightOptions,
+  ConversationAgentMemoryMaintenanceMode,
+  ConversationAgentOptions,
+  ConversationAgentRuntimeContext,
+  ConversationAgentRuntimeDefaults,
+  ConversationAgentRuntimeDefaultsInput,
+  ConversationAgentSendInput,
+  ConversationAgentSessionOptions,
+  ConversationAgentTurnResult,
+} from './headless/index.js';
+export {
+  QuickstartConversationCliRunnerService,
+  resolveQuickstartConversationCliDefaults,
+  runQuickstartConversationCli,
+} from './console/index.js';
+export type {
+  QuickstartConversationCliCredentialContext,
+  QuickstartConversationCliCredentialPreflightOptions,
+  QuickstartConversationCliLocalCommand,
+  QuickstartConversationCliLocalCommandContext,
+  QuickstartConversationCliMemoryMaintenanceMode,
+  QuickstartConversationCliRunnerDefaults,
+  QuickstartConversationCliRunnerDefaultsInput,
+  QuickstartConversationCliRunnerOptions,
+  QuickstartConversationCliTurnContext,
+} from './console/index.js';
